+++ base/deprecated.jl
@@ -1168,4 +1168,11 @@ for (dep, f, op) in [(:sumabs!, :sum!, :abs),
     end
 end
 
+# QuadGK moved to a package (#19741)
+function quadgk(args...)
+    error(string(quadgk, args, " has been moved to the package QuadGK.jl.\n",
+                 "Run Pkg.add(\"QuadGK\") to install QuadGK on Julia v0.6 and later, and then run `using QuadGK`."))
+end
+export quadgk
+
 # End deprecations scheduled for 0.6