+++ NEWS.md
@@ -57,6 +57,8 @@ This section lists changes that do not have deprecation warnings.
     The flip-side of this is that new method definitions should now reliably actually
     take effect, and be called when evaluating new code ([#265]).
 
+  * `quadgk` has been moved from Base into a separate package. ([#19741])
+
 Library improvements
 --------------------